style: Clang tidy support
* build: switch to run clang-format.py

* build: include test folder into clang-format range

* build: add .clang-format and .clang-tidy and revise style accordingly

* doc: after -03 compiler optim, new benchmark stats

+++ src/include/core/turtle_server.h
@@ -54,7 +54,7 @@ namespace TURTLE_SERVER {
 class TurtleServer {
  public:
   explicit TurtleServer(NetAddress server_address,
-                        int concurrency = std::thread::hardware_concurrency() -
+                        int concurrency = static_cast<int>(std::thread::hardware_concurrency()) -
                                           1)
       : pool_(std::make_unique<ThreadPool>(concurrency)),
         listener_(std::make_unique<Looper>()) {
